LPS-55084 Move HibernatePortalCacheManger out of LiferayEhcacheRegionFactory
